add demo controller and now images handle with storyboard
